fix: profile mw tests (#240)
* fix: profile mw tests

* rebase

+++ src/middlewares/base/Treasury.sol
@@ -44,7 +44,7 @@ contract Treasury is ITreasury {
         return _treasuryAddress;
     }
 
-    function getTreasuryFee() external view override returns (uint16) {
+    function getTreasuryFee() external view override returns (uint256) {
         return _treasuryFee;
     }
 }
